chore: move 'find_parent_bench' to utils
style: consistent spaces and formatting, code cleanup
